// generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    PrimaryFieldName: ComponentFramework.PropertyTypes.StringProperty;
    SearchFields: ComponentFramework.PropertyTypes.StringProperty;
    RelatedEntityName: ComponentFramework.PropertyTypes.StringProperty;
    RelationshipName: ComponentFramework.PropertyTypes.StringProperty;
    relatedRecords: ComponentFramework.PropertyTypes.DataSet;
}
export interface IOutputs {
}